more tests and ting

+++ src/Rafty/Concensus/AppendEntries.cs
@@ -6,8 +6,8 @@ namespace Rafty.Concensus
 
     public sealed class AppendEntries : Message
     {
-        public AppendEntries(long term, Guid leaderId, long previousLogIndex, long previousLogTerm,
-            List<LogEntry> entries, long leaderCommitIndex)
+        public AppendEntries(long term, Guid leaderId, int previousLogIndex, long previousLogTerm,
+            List<LogEntry> entries, int leaderCommitIndex)
             : base(Guid.NewGuid())
         {
             Term = term;
@@ -23,12 +23,12 @@ namespace Rafty.Concensus
         // leaderId so follower can redirect clients
         public Guid LeaderId { get; private set; }
         // prevLogIndex index of log entry immediately preceding new ones
-        public long PreviousLogIndex { get; private set; }
+        public int PreviousLogIndex { get; private set; }
         // prevLogTerm term of prevLogIndex entry
         public long PreviousLogTerm { get; private set; }
         // entries[] log entries to store (empty for heartbeat may send more than one for efficiency)
         public List<LogEntry> Entries { get; private set; }
         // leaderCommit leader’s commitIndex
-        public long LeaderCommitIndex { get; private set; }
+        public int LeaderCommitIndex { get; private set; }
     }
 }
