+++ src/Rafty/Log/LogEntry.cs
@@ -4,7 +4,7 @@
 
     public class LogEntry
     {
-        public LogEntry(string commandData, Type type, long term, long currentCommitIndex)
+        public LogEntry(string commandData, Type type, long term, int currentCommitIndex)
         {
             CommandData = commandData;
             Type = type;
@@ -15,6 +15,6 @@
         public string CommandData { get; private set; }
         public Type Type { get; private set; }
         public long Term { get; private set; }
-        public long CurrentCommitIndex { get; private set; }
+        public int CurrentCommitIndex { get; private set; }
     }
 }
